requirements.txt
streamlit
sentence-transformers
transformers
faiss-cpu
wikipedia-api
torch
scikit-learn